requirements.txt
streamlit==1.32.0
streamlit-lottie==0.0.5
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
python-dotenv==1.0.1
